refact/TripCompanionsForm: moves component within src/pages/trips/new/TripCompanionsForm/TripCompanionsForm.jsx
